Initial game-wide replacement of scrolling playfields
