+++ osu.Game.Rulesets.Mania/ManiaRuleset.cs
@@ -95,7 +95,6 @@ namespace osu.Game.Rulesets.Mania
                                 new ModCinema(),
                             },
                         },
-                        new ManiaModGravity()
                     };
 
                 default:
