+++ osu.Game.Rulesets.Taiko/Objects/Drawables/DrawableBarLine.cs
@@ -11,7 +11,7 @@ namespace osu.Game.Rulesets.Taiko.Objects.Drawables
     /// <summary>
     /// A line that scrolls alongside hit objects in the playfield and visualises control points.
     /// </summary>
-    public class DrawableBarLine : DrawableScrollingHitObject<TaikoHitObject>
+    public class DrawableBarLine : DrawableHitObject<TaikoHitObject>
     {
         /// <summary>
         /// The width of the line tracker.
