+++ osu.Game.Tests/osu.Game.Tests.csproj
@@ -142,7 +142,6 @@
     <Compile Include="Visual\TestCaseRoomInspector.cs" />
     <Compile Include="Visual\TestCaseScoreCounter.cs" />
     <Compile Include="Visual\TestCaseScrollingHitObjects.cs" />
-    <Compile Include="Visual\TestCaseScrollingPlayfield.cs" />
     <Compile Include="Visual\TestCaseSettings.cs" />
     <Compile Include="Visual\TestCaseSkipButton.cs" />
     <Compile Include="Visual\TestCaseSocial.cs" />
